+++ CoExpNetViz/src/main/java/be/ugent/psb/coexpnetviz/gui/jobinput/RunAnalysisTaskController.java
@@ -1,4 +1,4 @@
-package be.ugent.psb.coexpnetviz.gui;
+package be.ugent.psb.coexpnetviz.gui.jobinput;
 
 /*
  * #%L
