+++ CoExpNetViz/src/main/java/be/ugent/psb/coexpnetviz/io/SettingsIO.java
@@ -1,5 +1,27 @@
 package be.ugent.psb.coexpnetviz.io;
 
+/*
+ * #%L
+ * CoExpNetViz
+ * %%
+ * Copyright (C) 2015 - 2016 PSB/UGent
+ * %%
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ * 
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Lesser Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Lesser Public
+ * License along with this program.  If not, see
+ * <http://www.gnu.org/licenses/lgpl-3.0.html>.
+ * #L%
+ */
+
 import static java.nio.file.StandardOpenOption.CREATE;
 import static java.nio.file.StandardOpenOption.TRUNCATE_EXISTING;
 
@@ -18,30 +40,7 @@ import java.util.Map;
 import org.yaml.snakeyaml.Yaml;
 
 import be.ugent.psb.coexpnetviz.Context;
-
-/*
- * #%L
- * CoExpNetViz
- * %%
- * Copyright (C) 2015 PSB/UGent
- * %%
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- * 
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Lesser Public License for more details.
- * 
- * You should have received a copy of the GNU General Lesser Public
- * License along with this program.  If not, see
- * <http://www.gnu.org/licenses/lgpl-3.0.html>.
- * #L%
- */
-
-import be.ugent.psb.coexpnetviz.gui.model.JobInputModel;
+import be.ugent.psb.coexpnetviz.gui.jobinput.JobInputModel;
 
 /**
  * Read/write of CoExpNetViz settings
